# swagger/swagger.yaml
swagger: '2.0'
info:
  description: Swagger API for SmartbeeExam1.
  version: 1.0.0
  title: SmartbeeExam1
host: localhost
basePath: /v1
tags:
  - name: company
    description: Operations about company
  - name: client
    description: Operations about client
schemes:
  - https
  - http
paths:
  /company:
    post:
      tags:
        - company
      summary: Create company
      operationId: createCompany
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: array
            items:
              $ref: '#/definitions/Company'
      responses:
        default:
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Company'
    get:
      tags:
        - company
      summary: List all company
      operationId: listCompany
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Company'
        '400':
          description: Invalid company id supplied
        '404':
          description: Company not found
  '/company/{id}':
    get:
      tags:
        - company
      summary: Get company by company id
      operationId: readCompany
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Company'
        '400':
          description: Invalid company id supplied
        '404':
          description: Company not found
    put:
      tags:
        - company
      summary: Updated company
      operationId: updateCompany
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: string
        - in: body
          name: body
          description: Updated company object
          required: true
          schema:
            $ref: '#/definitions/Company'
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Company'
        '400':
          description: Invalid company supplied
        '404':
          description: Company not found
    delete:
      tags:
        - company
      summary: Delete company
      operationId: deleteCompany
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        '400':
          description: Invalid company id supplied
        '404':
          description: Company not found
  /client:
    post:
      tags:
        - client
      summary: Create client
      operationId: createClient
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: array
            items:
              $ref: '#/definitions/Client'
      responses:
        default:
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Client'
    get:
      tags:
        - client
      summary: List all client
      operationId: listClient
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Client'
        '400':
          description: Invalid client id supplied
        '404':
          description: Client not found
  '/client/{id}':
    get:
      tags:
        - client
      summary: Get client by client id
      operationId: readClient
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Client'
        '400':
          description: Invalid client id supplied
        '404':
          description: Client not found
    put:
      tags:
        - client
      summary: Updated client
      operationId: updateClient
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: string
        - in: body
          name: body
          description: Updated client object
          required: true
          schema:
            $ref: '#/definitions/Client'
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Client'
        '400':
          description: Invalid client supplied
        '404':
          description: Client not found
    delete:
      tags:
        - client
      summary: Delete client
      operationId: deleteClient
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        '400':
          description: Invalid client id supplied
        '404':
          description: Client not found
securityDefinitions:
  petstore_auth:
    type: oauth2
    authorizationUrl: 'http://petstore.swagger.io/oauth/dialog'
    flow: implicit
    scopes:
      'write:pets': modify pets in your account
      'read:pets': read your pets
  api_key:
    type: apiKey
    name: api_key
    in: header
definitions:
  Company:
    type: object
    properties:
      id:
        type: integer
        format: int32
      name:
        type: string
      address:
        type: string
      created_by:
        type: string
      created_at:
        type: integer
        format: int64
      updated_by:
        type: string
      updated_at:
        type: integer
        format: int64
    xml:
      name: Company
  Client:
    type: object
    properties:
      id:
        type: integer
        format: int32
      company_id:
        type: integer
        format: int32
      name:
        type: string
      email:
        type: string
      phone:
        type: string
      created_by:
        type: string
      created_at:
        type: integer
        format: int64
      updated_by:
        type: string
      updated_at:
        type: integer
        format: int64
    xml:
      name: Client